requirements.txt
pandas == 2.2.3
pillow == 11.1.0